pyproject.toml
[tool.poetry]
name = "vyper-teste"
version = "0.1.0"
description = ""
authors = ["Allan Assis <allan-bdc@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
vyper-config = "^1.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
